requirements.txt
dspy-ai
dspy
openai
python-dotenv
datasets
pandas